feat(tracking): track segmentID if cookie is present
This change detects if the `segmentID` cookie is present, and, if so, adds it to the o-tracking context object, so that it is tracked for all events on the page. segmentID is set as a session cookie, and is not always present.

segmentID is used for campaign tracking for inbound ads, through - for example - facebook ads. We need to track segmentID so we can attach anonymous user journeys to particular campaigns.