refactor(segmentId): switch to using query param over cookie
After talking with @commuterjoy about the use of cookies, and how spoor will track
context across visits, it was decided that this should use the query parameter over
a cookie, as we should minimise the amount of cookies that users have - especially
when it can be done another way. This change switches from the use of cookie detection
for segmentId to using the query param instead.

This requires https://github.com/Financial-Times/ft.com-cdn/issues/199 to be closed,
before going live.