Use data-trackable string as a component id
When browsing the codebase, it’s not clear that HTML `id`s are being used for more than just n-section (which is in the process of being removed). A few of us have been gradually stripping out these `id`s on that basis. It seems likely that this mistake will be made again, hence this change (where most `id`s have the same value as their respective `data-trackable` strings, and any discrepancies will be resolved shortly).

@adgad