Fix `make deploy` target (#1393)
If a target is overridden in a project's `Makefile`, then it must have content (i.e. run a command) in order for the original `n-gage` target not to run. As the `deploy` target in this project was only calling other targets and not defining its own content, `n-gage`'s `deploy` target was being run afterwards and then producing an error as it is only intended for deploying our apps to Heroku.