define tracking properties in a single object rather than in multiple variables
this is the first step to moving tracking for barriers out of n-ui. This moves the computation of tracking values into next-product and sends them as one object to n-ui for attaching to an o-tracking event. Additional work is required at a later stage to send this event directly from next-product but is not being done right now as it involves updating tracking code for the standalone barrier in the epaper repo
 🐿 v2.12.3